Update Identity.md
Changed the frame of the paragraph on national secruity and credit scoring agencies. I would actually prefer if they were pulled out and discussed as quite separate things. In Domains of Identity - they are in very different locations on the "map" one is legally regulated (credit scoring agencies) and relies on private sector data to create dossiers and sells that to the private sector.  Government surveillance and profiling of people can be legally mandated/accepted or be in the unlawful surveillance realm. These two "feel" a bit the same but they are quite different - if we don't like them we need to solve for each differnetly. I wrote domains of identity in part so folks could get clear about these important distictions.