Update 04-01-identity-and-personhood.md
"decentralized identity" =/= DID
Decentralized Identifiers (the protocol at W3C) are DIDs

The movement to created decentralized identity system is not the same as one of the protocols within that suite of protocols. 

I also don't think Kim was a big fan of DID's per-se he did overall back the intentions of the work of the movement to make decentralized identity protocols and the systems that use those protocols. 

Verifiable Credentials bare strong similarity to  Kim's key work CardSpace (also early on called InfoCard https://en.wikipedia.org/wiki/Windows_CardSpace) but which basically came out "to early" in the market to be adopted.