Update 04-01-identity-and-personhood.md
I corrected some of the language and made it clear there are two core standards for decentralized identity

Decentralized Identifiers and Verifiable Credentials along with ultra brief descriptions of what they do. 

I also streamlined the language in the next paragraph to take out the use of the word acronym DID for "decentralized identity".