Update 04-01-identity-and-personhood.md
Given that Adhaar is highlighted more prominently in the latest version of the text and MOSIP is given a whole paragraph. 

It felt like the adoption of the three party model ( holder, issuer, verifier) that is being deployed in the real world by the EU, USCIS and Bhutan (with others in the pipeline that are not public) was important to give space to.  

I also removed the paragraph about covid creds in Europe - the work to build out eIDAS2.0 supersedes that.