Update 03-01-living-in-a-plural-world.md
I added a foot note about Assemblage Theory. 
I mentioned this to Glen and Audrey many weeks ago. It was suggested I might contribute edits that folded it in. It seemed like the best way for it to be included and not burden the text trying explain/justify it was to add a foot note. 
I think Assemblage Theory is deeply pluralist and referencing it helps your argument that what you are saying is not "out there" but echoed in it (and many of the other things you name on this page).