Update 04-00-rights-os-and-digital-freedom.md
I added a small amount of material about actual signals that are promising - we don't have to "invent all the things" that are missing - because early seeds of many of them do exist. 
They are NOT widely adopted. And hopefully this will book can help accelerate the development of real world tools that leverage them. 

Addressing this issue. https://github.com/pluralitybook/plurality/issues/235